modify ttsapi.js to support .wav file
This is to support usage on a Roku player using the roAudioResource which only supports .wav files. This is the only method that will allow simultaneous playback of speech on a Roku while music is playing without interrupting playback. In addition I intend to add support for multiple command-line text to speech synthesizers.